✨ feat : Location 데이터 분리
